split out as i18n util
